Backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
requests==2.32.3
google-generativeai==0.8.4
pydantic==2.10.6
python-dotenv==1.0.1
